+++ menrva/app/src/main/cpp/abstracts/fft_interface_base.h
@@ -26,18 +26,18 @@
 
 class FftInterfaceBase {
 public:
-    FftInterfaceBase(unsigned int signalSize = 0, unsigned int componentSize = 0);
-    virtual int Initialize(unsigned int signalSize, unsigned int componentSize = 0);
+    FftInterfaceBase(size_t signalSize = 0, size_t componentSize = 0);
+    virtual size_t Initialize(size_t signalSize, size_t componentSize = 0);
     virtual void SignalToComponents(AudioBuffer* signal, AudioComponentsBuffer* components) = 0;
     virtual void ComponentsToSignal(AudioComponentsBuffer* components, AudioBuffer* signal) = 0;
     virtual sample* Allocate(size_t size) = 0;
     virtual void Deallocate(sample* data) = 0;
-    int GetSignalSize();
-    int GetComponentSize();
+    size_t GetSignalSize();
+    size_t GetComponentSize();
 
 protected:
-    int _SignalSize,
-        _ComponentSize;
+    size_t _SignalSize,
+           _ComponentSize;
 };
 
 #endif //MENRVA_FFT_INTERFACE_BASE_H
