+++ menrva/app/src/main/cpp/fft/fftw_interface.h
@@ -28,8 +28,8 @@ typedef std::map<std::string, fftw_plan_pair> PlanCache;
 
 class FftwInterface : public FftInterfaceBase {
 public:
-    FftwInterface(unsigned int signalSize = 0, unsigned int componentSize = 0);
-    int Initialize(unsigned int signalSize, unsigned int componentSize = 0) override;
+    FftwInterface(size_t signalSize = 0, size_t componentSize = 0);
+    size_t Initialize(size_t signalSize, size_t componentSize = 0) override;
     void SignalToComponents(AudioBuffer* signal, AudioComponentsBuffer* components) override;
     void ComponentsToSignal(AudioComponentsBuffer* components, AudioBuffer* signal) override;
     sample* Allocate(size_t size) override;
